+++ docs/Whats-New.md
@@ -267,6 +267,7 @@ New:
 - Customizable debris & meteor impact and warhead detonation behaviour (by Starkku, Otamaa)
 - Custom warhead debris animations (by Starkku)
 - Multiple burst shots / burst delay within infantry firing sequence (by Starkku)
+- Attached particle system for animations (by Starkku)
 
 Vanilla fixes:
 - Allow AI to repair structures built from base nodes/trigger action 125/SW delivery in single player missions (by Trsdy)